Disable $compile_check for Smarty
"If $compile_check is enabled (default), every template file and config file that is involved with the cache file is checked for modification. If any of the files have been modified since the cache was generated, the cache is immediately regenerated. This is a computational overhead, so for optimum performance set $compile_check to FALSE."